backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
openai
python-multipart
firebase-admin
pyjwt
email-validator 